Parse commands and increment a speed value
Creates a new function, alters the OnMessageHandler function, and
creates a global varible for speed

parse: Takes in a message as a parameter and does a super simple parse
on it. At this point if "Go" was typed into the twitch chat a value of 1
is returned, if "Stop" then a value of 0 is returned, and if it is
neithier a value of -1 is returned

OnMessageHandler: No longer has a consol.log but rather calls the parse
function and uses the returned value to eithier incrment or decrement
the arbitrary speed value. Also if it gets a -1 returned it knows that
the message recived was not a command