Create and save a new oauth token
creates a catch function called refresh to the client.connect() action.

refresh: reads the url that generate the new twitch token using the
refreshtoken. Then it parse the text from the URL and grabs the new
oauth token. Lastly it saves the new token into the config file and
connects the client.