Move video bitrate, size, and filter to config (#63)
This adds a configurable bitrate, resolution, and filter string to the config. A user should now have the ability to adjust their stream to suit different input devices. We mentioned in the readme that a user can adjust the video properties for their stream in the config file, but we didn't actually make that fully possible until now.